Separated logging in and signing up from user's programs, added exception handling
